+++ src/xembed.hh
@@ -6,7 +6,7 @@
 class Xembed : public fltk::Window {
    public:
       Xembed(int _w, int _h) : fltk::Window(_w, _h) {};
-      void embed(unsigned long xid);
+      void embed(uint32_t xid);
 };
 
 #endif
